ci: composer update, upgrade to php 8.2 and ajuste vite.config.mjs for https
This commit updates the composer packages to the latest version and php from 8.1 to 8.2. It also updates the vite server configuration for working with a Secure Development Server with the latest change on composer package. With these last update and using Laravel Herd, the Laravel Vite plugin will automatically detect and use the generated TLS certificate.